requirements.txt
redis==2.10.5
dill==0.2.5
RPi.GPIO==0.6.2
mock==2.0.0
